# docs/swagger.yaml
definitions:
  model.ImageFilter:
    properties:
      block_id:
        type: string
      id:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  model.ImageLimitResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.ImageResponse'
        type: array
      total:
        type: integer
    type: object
  model.ImageResponse:
    properties:
      block_id:
        type: string
      create_at:
        type: string
      data:
        type: string
      error:
        type: string
      id:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  model.ImageUpdate:
    properties:
      data:
        type: string
    type: object
info:
  contact: {}
  license: {}
  title: OCR BACKEND API
  version: "1.0"
paths:
  /api/v1/auth/excel:
    get:
      description: export excel, save file to *.xlsx
      parameters:
      - in: query
        name: block_id
        type: string
      - in: query
        name: id
        type: string
      - in: query
        name: status
        type: string
      - in: query
        name: user_id
        type: string
      - description: '''Bearer '' + token'
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ImageLimitResponse'
      summary: excel *.xlsx
      tags:
      - Excel
  /api/v1/auth/image/{image_id}:
    get:
      description: get list image
      parameters:
      - description: image id
        in: path
        name: image_id
        required: true
        type: string
      - description: '''Bearer '' + token'
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ImageResponse'
      summary: image
      tags:
      - Images
    post:
      description: update image
      parameters:
      - description: image id
        in: path
        name: image_id
        required: true
        type: string
      - description: '''Bearer '' + token'
        in: header
        name: Authorization
        required: true
        type: string
      - description: image content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ImageUpdate'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: image
      tags:
      - Images
  /api/v1/auth/images:
    delete:
      description: get list image
      parameters:
      - in: query
        name: block_id
        type: string
      - in: query
        name: id
        type: string
      - in: query
        name: status
        type: string
      - in: query
        name: user_id
        type: string
      - description: '''Bearer '' + token'
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: image
      tags:
      - Images
    get:
      description: get list image
      parameters:
      - in: query
        name: block_id
        type: string
      - in: query
        name: id
        type: string
      - in: query
        name: status
        type: string
      - in: query
        name: user_id
        type: string
      - description: number
        in: query
        name: limit
        type: string
      - description: number
        in: query
        name: page
        type: string
      - description: '''Bearer '' + token'
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ImageLimitResponse'
      summary: image
      tags:
      - Images
  /api/v1/auth/images/block-ids:
    get:
      description: get list image
      parameters:
      - description: '''Bearer '' + token'
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: image
      tags:
      - Images
  /api/v1/auth/object:
    post:
      description: upload object
      parameters:
      - description: add block id
        in: header
        name: block_id
        type: string
      - description: '''Bearer '' + token'
        in: header
        name: Authorization
        required: true
        type: string
      - description: add file multipart/form-data
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ImageResponse'
        "400":
          description: some info
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: '...'
          schema:
            type: string
      summary: upload object
      tags:
      - Object
  /api/v1/auth/object/{id}:
    get:
      description: download object
      parameters:
      - description: object id
        in: path
        name: id
        required: true
        type: string
      - description: '''Bearer '' + token'
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "400":
          description: some info
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: '...'
          schema:
            type: string
      summary: download object
      tags:
      - Object
  /api/v1/object/{id}:
    get:
      description: download object
      parameters:
      - description: object id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "400":
          description: some info
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: '...'
          schema:
            type: string
      summary: download object
      tags:
      - Object
swagger: "2.0"